+++ PostHog/PostHogStorage.swift
@@ -212,6 +212,8 @@ class PostHogStorage {
         case remoteConfig = "posthog.remoteConfig"
         case surveySeen = "posthog.surveySeen"
         case requestId = "posthog.requestId"
+        case personPropertiesForFlags = "posthog.personPropertiesForFlags"
+        case groupPropertiesForFlags = "posthog.groupPropertiesForFlags"
     }
 
     // The location for storing data that we always want to keep
@@ -369,6 +371,8 @@ class PostHogStorage {
         deleteSafely(url(forKey: .remoteConfig))
         deleteSafely(url(forKey: .surveySeen))
         deleteSafely(url(forKey: .requestId))
+        deleteSafely(url(forKey: .personPropertiesForFlags))
+        deleteSafely(url(forKey: .groupPropertiesForFlags))
     }
 
     func remove(key: StorageKey) {